Bug/check imds version 57 (#60)
* Adde check for IMDS version

* move metadata below endpoint declaration

Moved check for IMDS version to below the metadata_endpoint declaration to be able to take advantage of it.

* updated default to IMDSv2

updated the comments to reflect IMDSv2 as the default.

* Add IMDS version 2 to README.md

Changing IMDSv2 to be default.  Updating the README to reflect this change.

* dealing with IMDS being blocked

Taking in to account if a firewall is configured to block IMDS calls.  And setting default to version 2.

* check IMDSv2 Token endpoint

Checks the IMDSv2 Token endpoint to verify if IMDSv2 is active and available.
If not, checks to verify if IMDSv1 endpoint is active and available.

* wrap curl in variable

curl command failure cause script to abort. Wrapping inside a variable should allow the script to continue.

* adding sudo to curl

curl to IMDS endpoint fails as a normal user in some cases (including testing with docker-compose).

* identify version failure in output

Identify in the output if it was IMDSv1 or IMDSv2 that failed.

* remove version lock from Dockerfile

cnf-lint and pyrsistent unlocking version

* unlock cnf-lint and pyrsistent versions

Unlocking versions for cnf-lint and pyrsistent in Docker file for Ubuntu 20.04

* adding check for user override to IMDS version

Adding if statement to check if the IMDS version has been requested.