Bug/fix-if-statement-67 (#68)
* fixing unbound variable

Using -z in an if statement causes an unbound variable error if set -eu is declared but the variable has not been declared.  Changing thing to -n which checks if the variable is declared and/or undefined.
undefined is an empty string like bob=""
undeclared is if it has never been declared at all

* Check for IMDS Endpoint

This checks of the IMDS Endpoint is available.  If not if fails.

* Add error message when IMDS Endpoint is unavailable

Add an error message if the IMDS Metadata Endpoint is not available.